Create separate package for media browser implementation
