+++ app/src/main/java/com/marverenic/music/data/inject/JockeyGraph.java
@@ -1,8 +1,8 @@
 package com.marverenic.music.data.inject;
 
-import com.marverenic.music.player.JockeyBrowserService;
 import com.marverenic.music.player.MusicPlayer;
 import com.marverenic.music.player.ServicePlayerController;
+import com.marverenic.music.player.browser.JockeyBrowserService;
 import com.marverenic.music.ui.BaseActivity;
 import com.marverenic.music.ui.BaseLibraryActivity;
 import com.marverenic.music.ui.about.AboutActivity;
